requirements.txt
streamlit
openai
requests
pandas
streamlit-extras
python-dotenv